requirements.txt
certifi==2019.3.9
cycler==0.10.0
kiwisolver==1.0.1
matplotlib==3.0.3
mkl-fft==1.0.0
mkl-random==1.0.1
numpy==1.14.2
opencv-python==3.4.2.17
py-getch==1.0.1
pyparsing==2.4.0
pyserial==3.4
python-dateutil==2.8.0
pytz==2019.1
six==1.12.0
spinnaker-python==1.20.0.15
tornado==6.0.2
wincertstore==0.2